refactor(simple_server): refactor done. Compiles and runs.
